refactor routing to /devices and add db stub for integ tests
