+++ app/integration/_index.test.ts
@@ -3,6 +3,6 @@ import { test, expect } from "@playwright/test";
 test("redirects to devices list", async ({ page }) => {
   await page.goto("/", { waitUntil: "load" });
   // Wait for the redirect to complete
-  await page.waitForURL("**/devices/list", { timeout: 5000 });
-  expect(page.url()).toBe(`http://localhost:50123/devices/list`);
+  await page.waitForURL("**/devices", { timeout: 5000 });
+  expect(page.url()).toBe(`http://localhost:50123/devices`);
 });
