+++ app/routes/devices.$id.$name.tsx
@@ -114,7 +114,7 @@ export function ErrorBoundary() {
         </div>
       </div>
 
-      <Link to="/devices/list" className="link">
+      <Link to="/devices" className="link">
         home
       </Link>
     </div>
