+++ package.json
@@ -77,6 +77,7 @@
     "playwright": "^1.45.3",
     "postcss": "^8.4.38",
     "prettier": "3.3.2",
+    "prisma": "^6.8.1",
     "tailwindcss": "^3.4.17",
     "typescript": "^5.5.3",
     "vite": "^6.3.5",
